Fixed bean.
git-svn-id: https://svn.jfire.org/svn/main/trunk/JFireDunning@17946 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/dunning/DunningConfig.java
@@ -196,7 +196,7 @@ implements Serializable
 		this.dunningConfigID = dunningConfigID;
 		this.dunningAutoMode = dunningAutoMode;
 		
-		this.defaultTermOfPaymentMSec = 31 * 24 * 60 * 60 * 1000;
+		this.defaultTermOfPaymentMSec = (long)(31 * 24 * 60 * 60 * 1000);
 		
 		this.name = new DunningConfigName(this);
 		this.description = new DunningConfigDescription(this);
